Various fixes and improvements
- Console panels now mimick those of traces, with filename, copy action and context menus; also consolidates both components' file boilerplate into a mixin for de-duplication
- Fixed PHPUnit reporter throwing a warning for incompatible - writeProgress signatures
- Improved accuracy when scanning package.json files for Jest testing scripts
- Attempt to add icon to Windows titlebar
- Attempt to fix macOS Catalina bold fonts being too bold
