+++ src/styles/variables.scss
@@ -100,7 +100,7 @@
     --font-family-sans-serif: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', Arial, sans-serif;
     --font-family-monospace: 'SFMono-Regular', 'Fira Mono', Consolas, 'Liberation Mono', Menlo, Courier, monospace;
 
-    --font-weight-bold: 900;
+    --font-weight-bold: 800;
     --font-weight-semibold: 600;
     --font-weight-normal: 400;
     --font-weight-light: 300;
